app/requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg[binary]==3.2.1
boto3==1.34.151

